Move deprecated servicer to diff file
