+++ mlserver/grpc/utils.py
@@ -1,6 +1,20 @@
-from typing import Dict, Tuple
+import grpc
+
+from typing import Callable, Dict, Tuple
+from fastapi import status
 
 from grpc import ServicerContext
+
+from .logging import logger
+from ..errors import MLServerError
+
+
+STATUS_CODE_MAPPING = {
+    status.HTTP_400_BAD_REQUEST: grpc.StatusCode.INVALID_ARGUMENT,
+    status.HTTP_404_NOT_FOUND: grpc.StatusCode.NOT_FOUND,
+    status.HTTP_422_UNPROCESSABLE_ENTITY: grpc.StatusCode.FAILED_PRECONDITION,
+    status.HTTP_500_INTERNAL_SERVER_ERROR: grpc.StatusCode.INTERNAL,
+}
 
 
 def to_headers(context: ServicerContext) -> Dict[str, str]:
@@ -14,3 +28,18 @@ def to_headers(context: ServicerContext) -> Dict[str, str]:
 
 def to_metadata(headers: Dict[str, str]) -> Tuple[Tuple[str, str], ...]:
     return tuple(headers.items())
+
+
+def _grpc_status_code(err: MLServerError):
+    return STATUS_CODE_MAPPING.get(err.status_code, grpc.StatusCode.UNKNOWN)
+
+
+def handle_mlserver_error(f: Callable):
+    async def _inner(self, request, context):
+        try:
+            return await f(self, request, context)
+        except MLServerError as err:
+            logger.error(err)
+            await context.abort(code=_grpc_status_code(err), details=str(err))
+
+    return _inner
